requirements.txt
playwright==1.51.0
pytest==8.3.4
pytest-xdist==3.6.1
pytest-rerunfailures==15.1
pytest-html==4.1.1
allure-pytest==2.13.5
python-dotenv==1.0.1
PyYAML==6.0.2